fix(vst): audit-driven RT-safety + lifetime hardening

+++ vst/Source/Engine/Rng.h
@@ -21,14 +21,22 @@ class Mulberry32
 public:
     explicit Mulberry32(std::uint32_t seed) : a(seed) {}
 
-    float next()
+    // Returns double to match the JS reference (Math.imul + division yields
+    // a Number = IEEE-754 double). The previous `static_cast<float>` form
+    // dropped ~25 bits of precision on every draw — fine for the existing
+    // cross-target test vectors (which compare booleans and chord PCs,
+    // not raw PRNG floats), but a future high-precision sample table
+    // pinned in `docs/ai/tonnetz-test-vectors.json::mulberry32.samples`
+    // would diverge bit-for-bit. Callers that need a float still get one
+    // via implicit narrowing.
+    double next()
     {
         a = a + 0x6D2B79F5u;
         std::uint32_t t = a;
         t = imul(t ^ (t >> 15), t | 1u);
         t = t ^ (t + imul(t ^ (t >> 7), t | 61u));
         const std::uint32_t out = t ^ (t >> 14);
-        return static_cast<float>(static_cast<double>(out) / 4294967296.0);
+        return static_cast<double>(out) / 4294967296.0;
     }
 
 private:
